Change main bower file to source instead of min
It is the convention to publish the source as the main file instead of the min version.

It's nice to bundle the min version but providing it by default causes problems with build pipelines that further process the file. For example, if you concat/uglify after, grunt-contrib-concat will complain that it needs the .map file. Depending on the build pipeline, that map file can not be available easily.

To be totally honest, I have no idea if this is a true convention or not, but I have 20+ dependencies in bower and this is the first time I see the min.js file as the main file and it crashed my build. :-)